Update from GitHub

+++ routes/install.php
@@ -19,46 +19,45 @@ use Illuminate\Support\Facades\Config;
 |
 */
 
-/*session_start();
-
-$route_install = md5(rand(100, 90000));
-
-if (isset($_SESSION['route_install'])) {
-    $route_install = $_SESSION['route_install'];
-}
-echo $route_install;*/
 Route::get('/install_', function () {
-    Artisan::call('key:generate');
-    Artisan::call('config:clear');
-    Artisan::call('cache:clear');
+    if (
+        isset($_GET['appKey'])
+        && $_GET['appKey'] == md5(env("APP_KEY"))
+    ) {
+        Artisan::call('key:generate');
+        Artisan::call('config:clear');
+        Artisan::call('cache:clear');
 
-    $tables = DB::select('SHOW TABLES');
+        $tables = DB::select('SHOW TABLES');
 
-    foreach ($tables as $table) {
-        $tableName = reset($table);
-        Schema::disableForeignKeyConstraints();
-        Schema::dropIfExists($tableName);
-        Schema::enableForeignKeyConstraints();
+        foreach ($tables as $table) {
+            $tableName = reset($table);
+            Schema::disableForeignKeyConstraints();
+            Schema::dropIfExists($tableName);
+            Schema::enableForeignKeyConstraints();
 
-        echo 'Delete table:' . $tableName . '<br>';
+            echo 'Delete table:' . $tableName . '<br>';
+        }
+
+        Artisan::call('migrate');
+        Artisan::call('route:clear');
+
+        $user = new \App\Models\User();
+        $userPassword = \Illuminate\Support\Str::random(10);
+        $user->password = \Illuminate\Support\Facades\Hash::make($userPassword);
+        $userLogin = 'root@'.request()->getHttpHost();
+        $user->email = $userLogin;
+        $user->name = 'Admin Root';
+        $user->save();
+
+        echo 'Your login: <b>' . $userLogin . '</b><br>';
+        echo 'Your password: <b>' . $userPassword . '</b><br>';
+        echo 'Front: <a target="_blank" href="/">Front</a><br>';
+        echo 'Admin Dashboard: <a target="_blank" href="/' . Config::get('ailara.admin_prefix') . '">'.Config::get('ailara.admin_prefix').'</a><br>';
+
+
+        return str_replace("\n", "<br>", Artisan::output());
     }
 
-    Artisan::call('migrate');
-    Artisan::call('route:clear');
-
-    $user = new \App\Models\User();
-    $userPassword = \Illuminate\Support\Str::random(10);
-    $user->password = \Illuminate\Support\Facades\Hash::make($userPassword);
-    $userLogin = 'root@'.request()->getHttpHost();
-    $user->email = $userLogin;
-    $user->name = 'Admin Root';
-    $user->save();
-
-    echo 'Your login: <b>' . $userLogin . '</b><br>';
-    echo 'Your password: <b>' . $userPassword . '</b><br>';
-    echo 'Front: <a target="_blank" href="/">Front</a><br>';
-    echo 'Admin Dashboard: <a target="_blank" href="/' . Config::get('ailara.admin_prefix') . '">'.Config::get('ailara.admin_prefix').'</a><br>';
-
-
-    return str_replace("\n", "<br>", Artisan::output());
+    return 'false';
 });
